backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.115.5
uvicorn[standard]==0.34.0
python-multipart==0.0.17

# Database - PostgreSQL
asyncpg==0.29.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.35
alembic==1.14.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
bcrypt==4.2.1

# Data validation
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# CORS and middleware
python-dateutil==2.9.0.post0

# Optional: AI/ML for chatbot (if using OpenAI or similar)
# openai==1.58.1
# anthropic==0.42.0
